Tradier work - unfinished
This is as far as I got after working all afternoon trying to figure out
the ins and outs of the tradier API.  We have to get a connection to a
web page which returns a string in xml format, and that string then has
to be parsed to get the data we need.  I have the functions written out
to get that data but right now when I run a test to get the info for
google stocks, it crashes on line 35 with "HTTPError: HTTP ERROR 401:
Unauthorized"

I'm assuming that means it doesn't like the token I am passing it.  That
probably has to do with how my tradier account is set up.  I am looking
into it.  This model is close to being done as far as lines of code that
have to be written but it's probably going to take a lot of little
tweaks to get it to do the tradier connection and xml parsing.